Move message definitions to separate file
