+++ nano/node/bootstrap/message_deserializer.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <nano/node/common.hpp>
+#include <nano/node/messages.hpp>
 #include <nano/node/socket.hpp>
 
 #include <boost/system/error_code.hpp>
