+++ nano/node/transport/transport.hpp
@@ -4,6 +4,7 @@
 #include <nano/lib/rate_limiting.hpp>
 #include <nano/lib/stats.hpp>
 #include <nano/node/common.hpp>
+#include <nano/node/messages.hpp>
 #include <nano/node/socket.hpp>
 
 #include <boost/asio/ip/network_v6.hpp>
